Invert ChatLayout to be above the actual Chat page
Currently we are including the chat layout as part of the chat page.
This isn't really how Layouts are intended to be used. This commit moves
the ChatLayout above the Chat Page, and uses a RouterView in order to
mount the ChatPage inside of it. This allows us to get rid of the
contact panel when we are within the Forum layout and instead use it for
topic selection and other configuration.