Add toast on post, and ensure it loads (#483)
The Forum page was not reloading after a post due to the filter within
the computer property. It caused it not to detect that the computed prop
was dependent on the underlying array length. This adds a toast, and
fixes the reload, so that users don't post multiple times thinking their
post did not go through.