Give topic hints to user when posting
People are entering lots of random topics and sometimes mistyping
existing ones. This commit switches the topic input on the CreatePost
screen to be a QSelect with all the known topics available. It also
allows for manually specifying a new topic, however it is preferable
to keep posts within known topics.