Don't redownload contacts on route change
Move contact fetching logic to the contacts store so that it can
check to see if a contact already exists. Additionally, don't try to
set the active chat if the contact already exists. This caused
loops due when changing routes quickly in some cases where the
active chat will bounce around between the contacts which were clicked
quickly in the chat list.