Refactor message item encoding and decoding to be their own functions
In order to implement message forwarding, we need to be able to
recursively encode message entries. This requires the functions for
encoding and decoding entries, to be have their own names so they can be
recursively called. This commit refactors those portions of the relay
client to be their own functions to enable this future functionality.