Ensure seed is saved permanently and not regenerated before setup
Currently, users are losing funds due to clicking back after funding
their first wallet. This is because the seed regenerates and isn't
saved in the store.

This commit ensures the seed is generated on Setup load only if it does
not exist. Seeds on the "new character" page are also not regenerated on
load. And the wallet restore function now maintains the seed even when
the xpubkey is not yet generated.