Fix stealth transaction sends (#577)
During receipt of a stealth lotus send, the relay client was incorrectly
adding the stealth outgoing utxos to the local wallet. This would bust
sends by attempting to include those UTXOs in the inputs of a new
transaction. This commit adds back the missing "continue" which short
circuits the wallet add for outbound transactions.