First iteration of topic functionality to replace forum
This commit replaces the "Forum" page with the ability to
subscribe to a specific topic as a separate view. It also simplifies
what is required to make a post. This turns the "forum" more into
a group-chat like system; but with end-user moderation based local
filters on the amount burned for the message.