Remove nested QLayouts (#582)
Previously we used nested QLayouts early on in development because we
didn't understand out nested RouterViews worked. These "container"
versions of QLayout come with a lot of problems that caused rendering
issues. This commit removes them and adjusts where QPageContainer and
QPage are placed so that the nested layouts are no longer necessary.
This should speed up rendering, and hopefully fix the weird keyboard
issue on mobile.