Fix bug around updating the stored active chat (#586)
Update activeChatAddr from within the router, and ensure it is set to
null when the current route is no longer a chat. This fixes a bug where
in some cases a non-chat route was selected and the user would not be
able to navigate back to the appropriate user due to `acitveChatAddr`
still being set to that contact.