Ensure rehydrating topic messages share object w/ messageIndex (#594)
Reactivity on voting and other message actions in topics was for
messages which had been loaded prior to the last time the app had been
loaded. This was because the individual topic message lists were not
shared with the messageIndex. It is important that these objects are the
same otherwise both will need to be updated in order to trigger
reactivity. This commit changes the saved state to only store
payloadDigests for the topics, and then rehydrates them from the same
objects as the messageIndex.