Move most of the setup code to App.vue
Currently, most of the setup and loading code is on the MainLayout which
causes problems as the MainLayout needs to support *every* subroute. The
result of this is that QPages need to be nexted in weird ways across
chats, topics, etc. in order to support footers and headers properly.

By moving the setup code to App.vue this will allow for not having to
use MainLayout as the layout for all pages and allow a cleaner setup
between topics and chats.