Add `useWallet` and `useRelayClient` for composition API
A lot of components require use of a mix of the composition and class
APIs in vue due to not having a way to get at the wallet without `this`.
Thus, they cannot be used from `setup()` functions. This commit adds a
singleton for these two objects so they can be used from within the
composition API.