Move topics to their own tab on the left drawer
Topics could end up having the impact of occluding DMs. This commit
reorganizes the left drawer to have a specific topic tab. Also, add in a
badge to the DM view so that it's clear if a message is waiting for you
there. Finally, It also cleans up the LeftDrawer a bit as there was a
bunch of unused code in the setup.