Fix signup issue when having only one UTXO available (#640)
The constructTransaction function was not properly ignoring frozen UTXOs
leading to the relay signup failing after the user's metadata had been
sent to the metadata server. This was due to it trying to use the same
UTXO over again for relay signup (which had just been spent).
Additionally, when posting metadata any failures were being consumed
leading to it improperly reporting errors.