[bug] Fix clearing of UTXOs when resetting key on setup (#657)
The deleted UTXO cache was being updated when the wallet was reset
this was causing UTXOs not to be re-added when they were scanned
again later. So, even though the balances were updated in the
Pinia wallet store, the UTXOs would not be available for use when
trying to send another transaction.