[bugfix] Handle issues around deleting UTXOs on 50x errors
In order to ensure local UTXO state matches the blockchain we excessively
sync UTXOs on message sends or receives. Sometimes this causes a UTXO
to be deleted twice (which is intended behavior) but certain code paths
would cause this to improperly fail and generate an exception. Other
upstream code would then fail in inappropriate ways.

Additionally, the retries on message sending not properly handling errors
and retries.

This commit should fix both of those issues.